+++ CHANGELOG.md
@@ -11,6 +11,18 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 #### Fixed
 #### Security
 
+## [Unreleased] - YYYY-MM-DD
+#### Added
+- Listing latest version details from NPM (see the `-n` flag for `ls`).
+- Bumping to modules accounting for the latest versions on NPM (see `bump -n`).
+- Added "dry run" flag to `bump` so that no files are saved (see `bump -d`).
+- Spinner when pulling settings from NPM.
+#### Changed
+#### Deprecated
+#### Removed
+#### Fixed
+#### Security
+
 
 ## [1.4.0] - 2017-04-03
 #### Added